requirements.txt
requests>=2.31.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0
facebook-sdk>=3.1.0
python-twitter>=3.5
instagrapi>=2.0.0
pytelegrambot>=4.2.0
pymongo>=4.5.0
sqlalchemy>=2.0.0
pandas>=2.1.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.9.0
flake8>=6.1.0
python-dateutil>=2.8.2
pytz>=2023.3
apscheduler>=3.10.0
loguru>=0.7.0
pydantic>=2.4.0
fastapi>=0.103.0
uvicorn>=0.23.0
aiohttp>=3.8.0
asyncio>=3.4.3